requirements.txt
docker
pyyaml
six